--- fighters/common/src/energy.rs
@@ -1,5 +1,0 @@
-mod control;
-
-pub fn install() {
-    control::install();
-}
