+++ fighters/common/src/lib.rs
@@ -25,10 +25,8 @@ pub mod agent_inits;
 pub mod param;
 mod effect;
 // pub mod command_inputs;
-mod energy;
 
 pub fn install() {
     status::install();
-    energy::install();
     effect::install();
 }
