+++ src/system.rs
@@ -9,6 +9,7 @@ mod fighterspecializer;
 mod engine;
 // mod effects;
 pub mod controller;
+mod energy;
 mod modules;
 mod one_frame;
 mod css;
@@ -24,6 +25,7 @@ pub fn install() {
     engine::install();
     // effects::install();
     controller::install();
+    energy::install();
     modules::install();
     one_frame::install();
     css::install();
